requirements.txt
# Core dependencies
openai>=1.0.0
ebooklib>=0.18.0
html2text>=2024.2.26
ffmpeg-python>=0.2.0
pydantic>=2.0.0
nltk>=3.8.1
mem0ai>=1.0.0  # Vector database client (PyPI name is mem0ai)

# Utilities
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
tqdm>=4.66.1
colorama>=0.4.6

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# TTS Dependencies
# Default `pip install -r` resolves torch from PyPI. On a GPU machine, prefer a
# CUDA wheel so Kokoro can use CUDA — see README (devlog). Re-running this
# file after a manual CUDA install should not pin you back to +cpu.
torch>=2.2.0  # install CUDA build separately per README (GPU)
kokoro>=0.3.0
soundfile==0.12.1
librosa==0.10.2
# Optional: legacy ElevenLabs (voice_registry / cloud); not required for audio_pipeline
# elevenlabs>=1.0.0